pyproject.toml
[project]
name = "aic1-hands-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-cloud-aiplatform>=1.137.0",
    "pillow>=12.1.1",
]
